+++ tests/functions/test_dependent_objects.py
@@ -78,6 +78,85 @@ class TestDependentObjects(TestCase):
         assert objects[3] in deps
 
 
+class TestDependentObjectsWithColumnAliases(TestCase):
+    def create_models(self):
+        class User(self.Base):
+            __tablename__ = 'user'
+            id = sa.Column(sa.Integer, primary_key=True)
+            first_name = sa.Column(sa.Unicode(255))
+            last_name = sa.Column(sa.Unicode(255))
+
+        class Article(self.Base):
+            __tablename__ = 'article'
+            id = sa.Column(sa.Integer, primary_key=True)
+            author_id = sa.Column(
+                '_author_id', sa.Integer, sa.ForeignKey('user.id')
+            )
+            owner_id = sa.Column(
+                '_owner_id',
+                sa.Integer, sa.ForeignKey('user.id', ondelete='SET NULL')
+            )
+
+            author = sa.orm.relationship(User, foreign_keys=[author_id])
+            owner = sa.orm.relationship(User, foreign_keys=[owner_id])
+
+        class BlogPost(self.Base):
+            __tablename__ = 'blog_post'
+            id = sa.Column(sa.Integer, primary_key=True)
+            owner_id = sa.Column(
+                '_owner_id',
+                sa.Integer, sa.ForeignKey('user.id', ondelete='CASCADE')
+            )
+
+            owner = sa.orm.relationship(User)
+
+        self.User = User
+        self.Article = Article
+        self.BlogPost = BlogPost
+
+    def test_returns_all_dependent_objects(self):
+        user = self.User(first_name=u'John')
+        articles = [
+            self.Article(author=user),
+            self.Article(),
+            self.Article(owner=user),
+            self.Article(author=user, owner=user)
+        ]
+        self.session.add_all(articles)
+        self.session.commit()
+
+        deps = list(dependent_objects(user))
+        assert len(deps) == 3
+        assert articles[0] in deps
+        assert articles[2] in deps
+        assert articles[3] in deps
+
+    def test_with_foreign_keys_parameter(self):
+        user = self.User(first_name=u'John')
+        objects = [
+            self.Article(author=user),
+            self.Article(),
+            self.Article(owner=user),
+            self.Article(author=user, owner=user),
+            self.BlogPost(owner=user)
+        ]
+        self.session.add_all(objects)
+        self.session.commit()
+
+        deps = list(
+            dependent_objects(
+                user,
+                (
+                    fk for fk in get_referencing_foreign_keys(self.User)
+                    if fk.ondelete == 'RESTRICT' or fk.ondelete is None
+                )
+            ).limit(5)
+        )
+        assert len(deps) == 2
+        assert objects[0] in deps
+        assert objects[3] in deps
+
+
 class TestDependentObjectsWithManyReferences(TestCase):
     def create_models(self):
         class User(self.Base):
@@ -191,7 +270,6 @@ class TestDependentObjectsWithSingleTableInheritance(TestCase):
             __mapper_args__ = {
                 'polymorphic_identity': u'blog_post'
             }
-
 
         self.Category = Category
         self.TextItem = TextItem