+++ CHANGES.rst
@@ -4,6 +4,69 @@ Changelog
 Here you can see the full list of changes between each SQLAlchemy-Utils release.
 
 
+0.27.12 (2014-12-xx)
+^^^^^^^^^^^^^^^^^^^^
+
+- Fixed PhoneNumber string coercion (#93)
+
+
+0.27.11 (2014-12-06)
+^^^^^^^^^^^^^^^^^^^^
+
+- Added loose typed column checking support for get_column_key
+- Made get_column_key throw UnmappedColumnError to be consistent with SQLAlchemy
+
+
+0.27.10 (2014-12-03)
+^^^^^^^^^^^^^^^^^^^^
+
+- Fixed column alias handling in dependent_objects
+
+
+0.27.9 (2014-12-01)
+^^^^^^^^^^^^^^^^^^^
+
+- Fixed aggregated decorator many-to-many relationship handling
+- Fixed aggregated column alias handling
+
+
+0.27.8 (2014-11-13)
+^^^^^^^^^^^^^^^^^^^
+
+- Added is_loaded utility function
+- Removed deprecated has_any_changes
+
+
+0.27.7 (2014-11-03)
+^^^^^^^^^^^^^^^^^^^
+
+- Added support for Column and ColumnEntity objects in get_mapper
+- Made make_order_by_deterministic add deterministic column more aggressively
+
+
+0.27.6 (2014-10-29)
+^^^^^^^^^^^^^^^^^^^
+
+- Fixed assert_max_length not working with non nullable columns
+- Add PostgreSQL < 9.2 support for drop_database
+
+
+0.27.5 (2014-10-24)
+^^^^^^^^^^^^^^^^^^^
+
+- Made assert_* functions automatically rollback session
+- Changed make_order_by_deterministic attach order by primary key for queries without order by
+- Fixed alias handling in has_unique_index
+- Fixed alias handling in has_index
+- Fixed alias handling in make_order_by_deterministic
+
+
+0.27.4 (2014-10-23)
+^^^^^^^^^^^^^^^^^^^
+
+- Added assert_non_nullable, assert_nullable and assert_max_length testing functions
+
+
 0.27.3 (2014-10-22)
 ^^^^^^^^^^^^^^^^^^^